+++ src/Toolkit/Toolkit/UI/Controls/BasemapGallery/BasemapGallery.Windows.cs
@@ -16,14 +16,15 @@
 
 #if !__IOS__ && !__ANDROID__
 using System;
-using Esri.ArcGISRuntime.Portal;
-using Esri.ArcGISRuntime.UI.Controls;
+using System.Collections.ObjectModel;
+using System.Collections.Specialized;
+using System.ComponentModel;
 using Esri.ArcGISRuntime.Mapping;
+using Esri.ArcGISRuntime.Portal;
 using System.Collections.Generic;
 #if NETFX_CORE
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Markup;
 #else
 using System.Windows;
 using System.Windows.Controls;
@@ -38,16 +39,9 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
     /// If connected to a GeoView, changing the basemap selection will change the connected Map or Scene's basemap.
     /// Only basemaps whose spatial reference matches the map or scene's spatial reference can be selected for display.
     /// </remarks>
-    [TemplatePart(Name = "PART_InnerListView", Type = typeof(ListView))]
-    public class BasemapGallery : Control
+    public partial class BasemapGallery : Control
     {
-        private ListView? _listView;
-        private readonly BasemapGalleryDataSource _controller;
-        private ItemsPanelTemplate? _listTemplate;
-        private ItemsPanelTemplate? _gridTemplate;
-
-        // Track currently applied style to avoid unnecessary re-styling of list view
-        private BasemapGalleryViewStyle _currentlyAppliedStyle = BasemapGalleryViewStyle.Automatic;
+        private readonly BasemapGalleryController _controller = new BasemapGalleryController();
 
         /// <summary>
         /// Initializes a new instance of the <see cref="BasemapGallery"/> class.
@@ -55,55 +49,9 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
         public BasemapGallery()
         {
             DefaultStyleKey = typeof(BasemapGallery);
-            _controller = new BasemapGalleryDataSource();
-            DataContext = this;
-            SizeChanged += BasemapGallery_SizeChanged;
-            _controller.PropertyChanged += _controller_PropertyChanged;
-            _controller.CollectionChanged += _controller_CollectionChanged;
-            _ = _controller.PopulateFromDefaultList();
-        }
-
-        private void _controller_CollectionChanged(object sender, System.Collections.Specialized.NotifyCollectionChangedEventArgs e)
-        {
-            if (e.Action == System.Collections.Specialized.NotifyCollectionChangedAction.Reset)
-            {
-                if (ListView.Items.Contains(_controller.SelectedBasemap))
-                {
-                    ListView.SelectedItem = _controller.SelectedBasemap;
-                }
-            }
-        }
-
-        private void _controller_PropertyChanged(object sender, System.ComponentModel.PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName == nameof(_controller.SelectedBasemap))
-            {
-                if (_controller.SelectedBasemap != null)
-                {
-                    BasemapSelected?.Invoke(this, _controller.SelectedBasemap);
-                }
-
-                ListView.SelectedItem = _controller.SelectedBasemap;
-            }
-        }
-
-        /// <inheritdoc />
-#if NETFX_CORE
-        protected override void OnApplyTemplate()
-#else
-        public override void OnApplyTemplate()
-#endif
-        {
-            base.OnApplyTemplate();
-
-            ListView = GetTemplateChild("PART_InnerListView") as ListView;
-
-            if (ListView != null)
-            {
-                ListView.ItemsSource = _controller;
-            }
-
-            SetNewStyle(ActualWidth);
+            SizeChanged += BasemapGallerySizeChanged;
+            AvailableBasemaps = new ObservableCollection<BasemapGalleryItem>();
+            _ = _controller.LoadFromDefaultPortal(this);
         }
 
         private ListView? ListView
@@ -115,36 +63,93 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
                 {
                     if (_listView != null)
                     {
-                        _listView.SelectionChanged -= ListSelectionChanged;
+                        _listView.SelectionChanged -= ListViewSelectionChanged;
                     }
 
                     _listView = value;
 
                     if (_listView != null)
                     {
-                        _listView.SelectionChanged += ListSelectionChanged;
+                        _controller.HandleListViewChanged(this);
+                        _listView.SelectionChanged += ListViewSelectionChanged;
                     }
                 }
             }
         }
 
-        private void ListSelectionChanged(object? sender, SelectionChangedEventArgs e)
+        private static void SelectedBasemapChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {
-            if (e.AddedItems.Count > 0 && e.AddedItems[0] is BasemapGalleryItem item)
+            if (d is BasemapGallery gallery)
             {
-                if (item.IsValid)
+                gallery._controller.HandleSelectedBasemapChanged(gallery);
+            }
+        }
+
+        private static void GeoModelChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
+        {
+            if (d is BasemapGallery gallery)
+            {
+                if (e.OldValue is GeoModel oldModel)
                 {
-                    _controller.SelectedBasemap = item;
-                    BasemapSelected?.Invoke(this, item);
+                    oldModel.PropertyChanged -= gallery.GeoModelPropertyChanged;
                 }
-                else if (sender is ListView lv)
+
+                gallery._controller.HandleGeoModelChanged(gallery);
+
+                if (e.NewValue is GeoModel newModel)
                 {
-                    lv.SelectedItem = null;
+                    newModel.PropertyChanged += gallery.GeoModelPropertyChanged;
                 }
             }
         }
 
-        public event EventHandler<BasemapGalleryItem>? BasemapSelected;
+        private void GeoModelPropertyChanged(object? sender, PropertyChangedEventArgs e)
+        {
+            _controller.HandleGeoModelPropertyChanged(this, e.PropertyName);
+        }
+
+        private void ListViewSelectionChanged(object? sender, SelectionChangedEventArgs e)
+        {
+            if (ListView == null)
+            {
+                return;
+            }
+
+            _controller.HandleListViewSelectionChanged(this, ListView.SelectedIndex);
+        }
+
+        private static void PortalChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
+        {
+            if (d is BasemapGallery gallery)
+            {
+                _ = gallery._controller.HandlePortalChanged(gallery);
+            }
+        }
+
+        private static void AvailableBasemapsChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
+        {
+            if (d is BasemapGallery gallery)
+            {
+                if (e.OldValue is ObservableCollection<BasemapGalleryItem> oldAvailableBasemaps)
+                {
+                    oldAvailableBasemaps.CollectionChanged -= gallery.AvailableBasemapsCollectionChanged;
+                }
+
+                gallery._controller.HandleAvailableBasemapsChanged(gallery);
+
+                if (e.NewValue is ObservableCollection<BasemapGalleryItem> newAvailableBasemaps)
+                {
+                    newAvailableBasemaps.CollectionChanged += gallery.AvailableBasemapsCollectionChanged;
+                }
+            }
+        }
+
+        private void AvailableBasemapsCollectionChanged(object? sender, NotifyCollectionChangedEventArgs e)
+        {
+            _controller.HandleAvailableBasemapsCollectionChanged(this, e);
+        }
+
+        #region Convenience Properties
 
         /// <summary>
         /// Gets or sets the portal to use for displaying basemaps.
@@ -165,214 +170,85 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
         }
 
         /// <summary>
-        /// Gets or sets the style used for the container used to display items when showing basemaps in a list.
+        /// Gets or sets the gallery of basemaps to show.
         /// </summary>
-        public Style? ListItemContainerStyle
+        /// <remarks>
+        /// When <see cref="Portal"/> is set, this collection will be overwritten.
+        /// </remarks>
+        public IList<BasemapGalleryItem>? AvailableBasemaps
         {
-            get => GetValue(ListItemContainerStyleProperty) as Style;
-            set => SetValue(ListItemContainerStyleProperty, value);
+            get => GetValue(AvailableBasemapsProperty) as IList<BasemapGalleryItem>;
+            set => SetValue(AvailableBasemapsProperty, value);
         }
 
         /// <summary>
-        /// Gets or sets the style used for the container used to display items when showing basemaps in a grid.
+        /// Gets or sets the selected basemap.
         /// </summary>
-        public Style? GridItemContainerStyle
+        /// <remarks>
+        /// This will be set to the gallery item that matches the basemap in the <see cref="GeoModel"/>.
+        /// </remarks>
+        public BasemapGalleryItem? SelectedBasemap
         {
-            get => GetValue(GridItemContainerStyleProperty) as Style;
-            set => SetValue(GridItemContainerStyleProperty, value);
+            get => GetValue(SelectedBasemapProperty) as BasemapGalleryItem;
+            set => SetValue(SelectedBasemapProperty, value);
         }
 
-        /// <summary>
-        /// Gets or sets the data template used to show basemaps in a list.
-        /// </summary>
-        public DataTemplate? ListItemTemplate
-        {
-            get => GetValue(ListItemTemplateProperty) as DataTemplate;
-            set => SetValue(ListItemTemplateProperty, value);
-        }
+        #endregion Convenience Properties
 
-        /// <summary>
-        /// Gets or sets the data template used to show basemaps in a grid.
-        /// </summary>
-        public DataTemplate? GridItemTemplate
-        {
-            get => GetValue(GridItemTemplateProperty) as DataTemplate;
-            set => SetValue(GridItemTemplateProperty, value);
-        }
-
-        /// <summary>
-        /// Gets or sets the view style for the gallery.
-        /// </summary>
-        public BasemapGalleryViewStyle GalleryViewStyle
-        {
-            get => (BasemapGalleryViewStyle)GetValue(GalleryViewStyleProperty);
-            set => SetValue(GalleryViewStyleProperty, value);
-        }
-
-        /// <summary>
-        /// Gets or sets the width threshold to use for deciding between grid and list views when <see cref="GalleryViewStyle"/> is <see cref="BasemapGalleryViewStyle.Automatic"/>.
-        /// </summary>
-        public double ViewStyleWidthThreshold
-        {
-            get => (double)GetValue(ViewStyleWidthThresholdProperty);
-            set => SetValue(ViewStyleWidthThresholdProperty, value);
-        }
-
-        public IEnumerable<BasemapGalleryItem>? OverrideList
-        {
-            get => GetValue(OverrideListProperty) as IEnumerable<BasemapGalleryItem>;
-            set => SetValue(OverrideListProperty, value);
-        }
-
-        // Using a DependencyProperty as the backing store for OverrideList.  This enables animation, styling, binding, etc...
-        public static readonly DependencyProperty OverrideListProperty =
-            DependencyProperty.Register("OverrideList", typeof(IEnumerable<BasemapGalleryItem>), typeof(BasemapGallery), new PropertyMetadata(null, OnOverrideListPropertyChanged));
-
-        private static void OnGeoModelPropertyChanged(DependencyObject d, DependencyPropertyChangedEventArgs e) => ((BasemapGallery)d)._controller.GeoModel = e.NewValue as GeoModel;
-
-        private static void OnPortalPropertyChanged(DependencyObject d, DependencyPropertyChangedEventArgs e) => ((BasemapGallery)d)._controller.PopulateBasemapsForPortal(e.NewValue as ArcGISPortal);
-
-        private static void OnOverrideListPropertyChanged(DependencyObject d, DependencyPropertyChangedEventArgs e) => ((BasemapGallery)d)._controller.SetOverrideList(e.NewValue as IEnumerable<BasemapGalleryItem>);
-
-        private static void OnViewLayoutPropertyChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            if (d is BasemapGallery gallery)
-            {
-                gallery.SetNewStyle(gallery.ActualWidth);
-            }
-        }
-
-        private static void OnStyleOrTemplateChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            if (d is BasemapGallery gallery)
-            {
-                gallery.UpdateListViewForStyle(gallery.GalleryViewStyle, true);
-            }
-        }
-
-        /// <summary>
-        /// Identifies the <see cref="ListItemContainerStyle"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty ListItemContainerStyleProperty =
-            DependencyProperty.Register(nameof(ListItemContainerStyle), typeof(Style), typeof(BasemapGallery), new PropertyMetadata(null, OnStyleOrTemplateChanged));
-
-        /// <summary>
-        /// Identifies the <see cref="GridItemContainerStyle"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty GridItemContainerStyleProperty =
-            DependencyProperty.Register(nameof(GridItemContainerStyle), typeof(Style), typeof(BasemapGallery), new PropertyMetadata(null, OnStyleOrTemplateChanged));
-
-        /// <summary>
-        /// Identifies the <see cref="ListItemTemplate"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty ListItemTemplateProperty =
-            DependencyProperty.Register(nameof(ListItemTemplate), typeof(DataTemplate), typeof(BasemapGallery), new PropertyMetadata(null, OnStyleOrTemplateChanged));
-
-        /// <summary>
-        /// Identifies the <see cref="GridItemTemplate"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty GridItemTemplateProperty =
-            DependencyProperty.Register(nameof(GridItemTemplate), typeof(DataTemplate), typeof(BasemapGallery), new PropertyMetadata(null, OnStyleOrTemplateChanged));
+        #region Dependency Properties
 
         /// <summary>
         /// Identifies the <see cref="GeoModel"/> dependency property.
         /// </summary>
         public static readonly DependencyProperty GeoModelProperty =
-            DependencyProperty.Register(nameof(GeoModel), typeof(GeoModel), typeof(BasemapGallery), new PropertyMetadata(null, OnGeoModelPropertyChanged));
+            DependencyProperty.Register(nameof(GeoModel), typeof(GeoModel), typeof(BasemapGallery), new PropertyMetadata(null, GeoModelChanged));
 
         /// <summary>
-        /// Identifies the <see cref="Portal"/> dependency proeprty.
+        /// Identifies the <see cref="Portal"/> dependency property.
         /// </summary>
         public static readonly DependencyProperty PortalProperty =
-            DependencyProperty.Register(nameof(Portal), typeof(ArcGISPortal), typeof(BasemapGallery), new PropertyMetadata(null, OnPortalPropertyChanged));
+            DependencyProperty.Register(nameof(Portal), typeof(ArcGISPortal), typeof(BasemapGallery), new PropertyMetadata(null, PortalChanged));
 
         /// <summary>
-        /// Identifies the <see cref="GalleryViewStyle"/> dependency property.
+        /// Identifies the <see cref="AvailableBasemaps"/> dependency property.
         /// </summary>
-        public static readonly DependencyProperty GalleryViewStyleProperty =
-            DependencyProperty.Register(nameof(GalleryViewStyle), typeof(BasemapGalleryViewStyle), typeof(BasemapGallery), new PropertyMetadata(BasemapGalleryViewStyle.Automatic, OnViewLayoutPropertyChanged));
+        public static readonly DependencyProperty AvailableBasemapsProperty =
+            DependencyProperty.Register(nameof(AvailableBasemaps), typeof(IList<BasemapGalleryItem>), typeof(BasemapGallery), new PropertyMetadata(null, AvailableBasemapsChanged));
 
         /// <summary>
-        /// Identifies the <see cref="ViewStyleWidthThreshold"/> dependency property.
+        /// Identifies the <see cref="SelectedBasemap"/> dependency property.
         /// </summary>
-        public static readonly DependencyProperty ViewStyleWidthThresholdProperty =
-            DependencyProperty.Register(nameof(ViewStyleWidthThreshold), typeof(double), typeof(BasemapGallery), new PropertyMetadata(440.0, OnViewLayoutPropertyChanged));
+        public static readonly DependencyProperty SelectedBasemapProperty =
+            DependencyProperty.Register(nameof(SelectedBasemap), typeof(BasemapGalleryItem), typeof(BasemapGallery), new PropertyMetadata(null, SelectedBasemapChanged));
+        #endregion Dependency Properties
 
-        private void BasemapGallery_SizeChanged(object sender, SizeChangedEventArgs e) => SetNewStyle(e.NewSize.Width);
+        /// <summary>
+        /// Event raised when a basemap is selected.
+        /// </summary>
+        public event EventHandler<BasemapGalleryItem>? BasemapSelected;
 
-        private void SetNewStyle(double currentSize)
+        #region Controller Callbacks
+        internal void SetListViewSource(IList<BasemapGalleryItem> newSource)
         {
-            if (ListView == null)
+            if (ListView != null)
             {
-                return;
-            }
-
-            if (GalleryViewStyle == BasemapGalleryViewStyle.Automatic)
-            {
-                UpdateListViewForStyle(currentSize >= ViewStyleWidthThreshold ? BasemapGalleryViewStyle.Grid : BasemapGalleryViewStyle.List);
-            }
-            else
-            {
-                UpdateListViewForStyle(GalleryViewStyle);
+                ListView.ItemsSource = newSource;
             }
         }
 
-        private void UpdateListViewForStyle(BasemapGalleryViewStyle style, bool forceReset = false)
+        internal void SetListViewSelection(BasemapGalleryItem? item)
         {
-            if (_currentlyAppliedStyle == style && !forceReset)
+            if (ListView != null)
             {
-                return;
+                ListView.SelectedItem = item;
             }
-
-            if (ListView == null)
-            {
-                return;
-            }
-
-            if (style == BasemapGalleryViewStyle.List)
-            {
-                ListView.ItemContainerStyle = ListItemContainerStyle;
-                ListView.ItemTemplate = ListItemTemplate;
-                ListView.ItemsPanel = _listTemplate ??= GetItemsPanelTemplate(typeof(StackPanel));
-            }
-            else if (style == BasemapGalleryViewStyle.Grid)
-            {
-                ListView.ItemContainerStyle = GridItemContainerStyle;
-                ListView.ItemTemplate = GridItemTemplate;
-                ListView.ItemsPanel = _gridTemplate ??= GetItemsPanelTemplate(typeof(WrapPanel));
-            }
-
-            _currentlyAppliedStyle = style;
         }
 
-#if NETFX_CORE
-        // Shim makes cross-platform code easier
-        private class WrapPanel
+        internal void NotifyBasemapSelected(BasemapGalleryItem item)
         {
+            BasemapSelected?.Invoke(this, item);
         }
-#endif
-
-        private static ItemsPanelTemplate? GetItemsPanelTemplate(Type panelType)
-        {
-#if !NETFX_CORE
-            return new ItemsPanelTemplate(new FrameworkElementFactory(panelType));
-#else
-            if (panelType == typeof(StackPanel))
-            {
-                string xaml = "<ItemsPanelTemplate xmlns='http://schemas.microsoft.com/winfx/2006/xaml/presentation' xmlns:x='http://schemas.microsoft.com/winfx/2006/xaml'>"
-                    + "<ItemsStackPanel />"
-                    + "</ItemsPanelTemplate>";
-                return XamlReader.Load(xaml) as ItemsPanelTemplate;
-            }
-            else
-            {
-                string xaml = "<ItemsPanelTemplate   xmlns='http://schemas.microsoft.com/winfx/2006/xaml/presentation' xmlns:x='http://schemas.microsoft.com/winfx/2006/xaml'>"
-                                   + "<ItemsWrapGrid Orientation=\"Horizontal\" />"
-                    + "</ItemsPanelTemplate>";
-                return XamlReader.Load(xaml) as ItemsPanelTemplate;
-            }
-#endif
-        }
+        #endregion Controller Callbacks
     }
 }
 #endif
